Move spec model classes own module
Move all the classes that implements spec data to a separate module that is
independent of pylons config import.  This opens the way to use the spec
model classes outside of the pylons application.  Now the use a factory
to generate model class that can point and any database the user wishes to use.